requirements.txt
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
tensorflow==2.15.0
keras==2.15.0
python-dotenv==1.0.0
schedule==1.2.1
requests==2.31.0
alpaca-py==0.13.1
ta==0.11.0
tqdm==4.66.1
joblib==1.3.2 